Test(WinningAmountTest): Move winning amount test from LottoTest to WinningAmountTest
